requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.6
django-braces==1.14.0
django-cors-headers==3.8.0
django-oauth-toolkit==1.5.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
idna==3.2
jwcrypto==1.0
oauthlib==3.1.1
Pillow==8.3.1
pycparser==2.20
PyMySQL==1.0.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.7
wrapt==1.12.1
